회고 API를 연동하라
